requirements.txt
numpy==1.24.1
pyarrow==10.0.1
pyelftools==0.29
python-magic==0.4.27
